+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - 2020-06-28
+### Added
+- Color Extensions to create color by changing one parameter of the color.
+- Helper to use css colors
 
-## [1.0.0] - 2022-06-27
+## [1.1.0] - 2022-06-27
 ### Breaking Changes
 - Requires LiteNinja-Utils
 ### Added